Restructure build
- Rename website project to docs
- Use same scalafmt version as IntelliJ
