+++ mdoc-docs/src/main/scala/mdoc/docs/MdocModifier.scala
@@ -1,4 +1,4 @@
-package mdoc.website
+package mdoc.docs
 
 import java.io.ByteArrayOutputStream
 import java.io.PrintStream
