+++ project/plugins.sbt
@@ -2,4 +2,3 @@ addSbtPlugin("io.get-coursier" % "sbt-coursier" % coursier.util.Properties.versi
 addSbtPlugin("com.eed3si9n" % "sbt-buildinfo" % "0.9.0")
 addSbtPlugin("com.eed3si9n" % "sbt-assembly" % "0.14.6")
 addSbtPlugin("com.geirsson" % "sbt-ci-release" % "1.2.1")
-addSbtPlugin("com.typesafe.sbt" % "sbt-ghpages" % "0.6.2")
